+++ components/TagList.php
@@ -5,7 +5,7 @@ namespace GinoPane\BlogTaxonomy\Components;
 use Cms\Classes\Page;
 use GinoPane\BlogTaxonomy\Plugin;
 use GinoPane\BlogTaxonomy\Models\Tag;
-use Illuminate\Database\Eloquent\Collection;
+use October\Rain\Database\Collection;
 
 /**
  * Class TagList
@@ -24,32 +24,46 @@ class TagList extends ComponentAbstract
     public $tags = [];
 
     /**
-     * Reference to the page name for linking to tags page
+     * Reference to the page name for linking to tag page
      *
      * @var string
      */
-    public $tagsPage;
+    private $tagPage;
+
+    /**
+     * Reference to the page name for linking to all tags page
+     *
+     * @var string
+     */
+    private $tagsPage;
+
+    /**
+     * URL to the page where all tags are listed
+     *
+     * @var string
+     */
+    public $tagsPageUrl;
 
     /**
      * If the tag list should be ordered by another attribute
      *
      * @var string
      */
-    public $orderBy;
+    private $orderBy;
 
     /**
-     * Whether display or not empty tags
+     * Whether display empty tags or not
      *
      * @var bool
      */
-    public $displayEmpty;
+    private $displayEmpty;
 
     /**
      * Filter tags for the post defined by slug
      *
      * @var string
      */
-    public $postSlug;
+    private $postSlug;
 
     /**
      * Limits the number of records to display
@@ -59,11 +73,58 @@ class TagList extends ComponentAbstract
     public $limit;
 
     /**
+     * Count whether overall amount of tags or amount of tags under "limit" only
+     *
+     * @var bool
+     */
+    private $exposeTotalCount;
+
+    /**
+     * Contains either the total number of tags in the list limited by "limit", or
+     * the total number of tags which may be more than the limit if "exposeTotalCount" is used
+     *
+     * @var int
+     */
+    public $totalCount;
+
+    /**
+     * Whether enable tag filter input or not
+     *
+     * @var bool
+     */
+    private $enableTagFilter;
+
+    /**
+     * Whether include tag filter input or not
+     *
+     * @var bool
+     */
+    public $tagFilterEnabled;
+
+    /**
+     * Tag filter options
+     */
+    const TAG_FILTER_NEVER = 'never';
+    const TAG_FILTER_ALWAYS = 'always';
+    const TAG_FILTER_ON_OVERFLOW = 'on_overflow';
+
+    /**
+     * Translations for tag filter options
+     *
+     * @var array
+     */
+    private static $enableTagFilterOptions = [
+        self::TAG_FILTER_NEVER => Plugin::LOCALIZATION_KEY . 'components.tag_list.tag_filter_options.never',
+        self::TAG_FILTER_ALWAYS => Plugin::LOCALIZATION_KEY . 'components.tag_list.tag_filter_options.always',
+        self::TAG_FILTER_ON_OVERFLOW => Plugin::LOCALIZATION_KEY . 'components.tag_list.tag_filter_options.on_overflow'
+    ];
+
+    /**
      * Component Registration
      *
      * @return  array
      */
-    public function componentDetails()
+    public function componentDetails(): array
     {
         return [
             'name'        => Plugin::LOCALIZATION_KEY . 'components.tag_list.name',
@@ -76,23 +137,16 @@ class TagList extends ComponentAbstract
      *
      * @return  array
      */
-    public function defineProperties()
+    public function defineProperties(): array
     {
         return [
+
+            //General properties
             'displayEmpty' => [
                 'title'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.display_empty_title',
                 'description'       => Plugin::LOCALIZATION_KEY . 'components.tag_list.display_empty_description',
                 'type'              => 'checkbox',
                 'default'           => false,
-                'showExternalParam' => false
-            ],
-            'limit' => [
-                'title'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_title',
-                'description'       => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_description',
-                'type'              => 'string',
-                'default'           => '0',
-                'validationPattern' => '^[0-9]+$',
-                'validationMessage' => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_validation_message',
                 'showExternalParam' => false
             ],
             'orderBy' => [
@@ -107,15 +161,60 @@ class TagList extends ComponentAbstract
                 'description' => Plugin::LOCALIZATION_KEY . 'components.tag_list.post_slug_description',
                 'type'        => 'string'
             ],
+
+            //Limit properties
+            'limit' => [
+                'group'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_group',
+                'title'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_title',
+                'description'       => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_description',
+                'type'              => 'string',
+                'default'           => '0',
+                'validationPattern' => '^[0-9]+$',
+                'validationMessage' => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_validation_message',
+                'showExternalParam' => false
+            ],
+            'exposeTotalCount' => [
+                'group'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_group',
+                'title'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.expose_total_count_title',
+                'description'       => Plugin::LOCALIZATION_KEY . 'components.tag_list.expose_total_count_description',
+                'type'              => 'checkbox',
+                'default'           => false,
+                'showExternalParam' => false
+            ],
+            'enableTagFilter' => [
+                'group'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.limit_group',
+                'title'             => Plugin::LOCALIZATION_KEY . 'components.tag_list.tag_filter_title',
+                'description'       => Plugin::LOCALIZATION_KEY . 'components.tag_list.tag_filter_description',
+                'type'              => 'dropdown',
+                'default'           => self::TAG_FILTER_NEVER,
+                'showExternalParam' => false
+            ],
+
+            //Links
+            'tagPage' => [
+                'title'         => Plugin::LOCALIZATION_KEY . 'components.tag_list.tag_page_title',
+                'group'         => Plugin::LOCALIZATION_KEY . 'components.post_list_abstract.links_group',
+                'description'   => Plugin::LOCALIZATION_KEY . 'components.tag_list.tag_page_description',
+                'type'          => 'dropdown',
+                'showExternalParam' => false
+            ],
+
             'tagsPage' => [
                 'title'         => Plugin::LOCALIZATION_KEY . 'components.tag_list.tags_page_title',
-                'group'         =>  Plugin::LOCALIZATION_KEY . 'components.post_list_abstract.links_group',
+                'group'         => Plugin::LOCALIZATION_KEY . 'components.post_list_abstract.links_group',
                 'description'   => Plugin::LOCALIZATION_KEY . 'components.tag_list.tags_page_description',
                 'type'          => 'dropdown',
-                'default'       => 'blog/tag',
                 'showExternalParam' => false
             ],
         ];
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getTagPageOptions()
+    {
+        return Page::sortBy('baseFileName')->lists('baseFileName', 'baseFileName');
     }
 
     /**
@@ -127,9 +226,9 @@ class TagList extends ComponentAbstract
     }
 
     /**
-     * @return mixed
+     * @return array
      */
-    public function getOrderByOptions()
+    public function getOrderByOptions(): array
     {
         $order = $this->translate(Tag::$sortingOptions);
 
@@ -139,25 +238,37 @@ class TagList extends ComponentAbstract
     }
 
     /**
+     * @return array
+     */
+    public function getEnableTagFilterOptions(): array
+    {
+        return $this->translate(self::$enableTagFilterOptions);
+    }
+
+    /**
      * Prepare and return a tag list
      *
      * @return void
      */
     public function onRun()
     {
-        $this->tagsPage = $this->getProperty('tagsPage');
+        $this->tagPage = (string) $this->getProperty('tagPage');
+        $this->tagsPage = (string) $this->getProperty('tagsPage');
+
         $this->orderBy = $this->getProperty('orderBy');
-        $this->postSlug = $this->property('postSlug');
-        $this->displayEmpty = $this->getProperty('displayEmpty');
-        $this->limit =  $this->getProperty('limit');
+        $this->postSlug = $this->getProperty('postSlug');
+        $this->displayEmpty = (bool) $this->getProperty('displayEmpty');
+        $this->limit =  (int) $this->getProperty('limit');
+        $this->exposeTotalCount =  (bool) $this->getProperty('exposeTotalCount');
+        $this->enableTagFilter = (string) $this->getProperty('enableTagFilter');
 
         $this->tags = $this->listTags();
     }
 
     /**
-     * @return mixed
+     * @return Collection
      */
-    protected function listTags()
+    private function listTags(): Collection
     {
         $tags = Tag::listFrontend([
             'sort' => $this->orderBy,
@@ -166,10 +277,59 @@ class TagList extends ComponentAbstract
             'post' => $this->postSlug
         ]);
 
-        $tagComponent = $this->getComponent(TagPosts::NAME, $this->tagsPage);
-
-        $this->setUrls($tags, $this->tagsPage, $this->controller, ['tag' => $this->urlProperty($tagComponent, 'tag')]);
+        $this->handleCount($tags);
+        $this->handleTagUrls($tags);
+        $this->handleTagFilter();
 
         return $tags;
+    }
+
+    /**
+     * @param $tags
+     */
+    private function handleTagUrls($tags)
+    {
+        $tagComponent = $this->getComponent(TagPosts::NAME, $this->tagPage);
+
+        $this->tagsPageUrl = $this->controller->pageUrl($this->tagsPage);
+
+        $this->setUrls($tags, $this->tagPage, $this->controller, ['tag' => $this->urlProperty($tagComponent, 'tag')]);
+    }
+
+    /**
+     * @param $tags
+     */
+    private function handleCount($tags)
+    {
+        $this->totalCount = $tags->count();
+
+        if ($this->exposeTotalCount && ($this->limit > 0) && ($this->totalCount === $this->limit)) {
+            $this->totalCount = Tag::listFrontend([
+                'displayEmpty' => $this->displayEmpty,
+                'post' => $this->postSlug
+            ])->count();
+        }
+    }
+
+    /**
+     * Enable tag filter input if required
+     */
+    private function handleTagFilter()
+    {
+        $this->tagFilterEnabled = false;
+
+        switch ($this->enableTagFilter) {
+            // fall-through is intentional here, setting of tagFilterEnabled to true is valid for both cases
+            case self::TAG_FILTER_ON_OVERFLOW:
+                if (!($this->totalCount > $this->limit)) {
+                    break;
+                }
+            case self::TAG_FILTER_ALWAYS:
+                $this->tagFilterEnabled = true;
+        }
+
+        if ($this->tagFilterEnabled) {
+            $this->addJs('/plugins/' . Plugin::DIRECTORY_KEY . '/assets/js/jquery.mark.min.js');
+        }
     }
 }